Fix deprecation warnings in test suite
Replace in-place ndarray shape assignment with np.reshape (deprecated in NumPy 2.5) and pass an explicit dtype to h5py create_dataset to preserve the prior float32 default.

+++ src/qcodes_loop/data/hdf5_format.py
@@ -285,7 +285,9 @@ class HDF5Format(Formatter):
             name = array.array_id
 
         # Create the hdf5 dataset
-        dset = group.create_dataset(array.array_id, (0, 1), maxshape=(None, 1))
+        dset = group.create_dataset(
+            array.array_id, (0, 1), maxshape=(None, 1), dtype="f4"
+        )
         dset.attrs["label"] = _encode_to_utf8(str(label))
         dset.attrs["name"] = _encode_to_utf8(str(name))
         dset.attrs["unit"] = _encode_to_utf8(str(array.unit or ""))
